Teacher & Course Card Code Optimization
